+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/tools/ModuleTools.java
@@ -130,7 +130,7 @@ public class ModuleTools
 
     super.onClientRegister(registry);
 
-    registry.registerItemModelStrategy(() -> {
+    registry.registerClientModelRegistrationStrategy(() -> {
 
       for (ItemWorktableTool item : ModuleTools.this.registeredToolList) {
         String resourcePath = item.getMaterial().isHighlighted() ? item.getName() + "_highlighted" : item.getName();