+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/block/BlockModelRegistrationStrategy.java
@@ -1,6 +1,6 @@
 package com.codetaylor.mc.artisanworktables.modules.worktables.block;
 
-import com.codetaylor.mc.athenaeum.registry.strategy.IModelRegistrationStrategy;
+import com.codetaylor.mc.athenaeum.registry.strategy.IClientModelRegistrationStrategy;
 import com.google.common.base.Preconditions;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
@@ -12,11 +12,11 @@ import net.minecraftforge.client.model.ModelLoader;
 import javax.annotation.Nonnull;
 
 class BlockModelRegistrationStrategy
-    implements IModelRegistrationStrategy {
+    implements IClientModelRegistrationStrategy {
 
   private Block block;
 
-  BlockModelRegistrationStrategy(Block block) {
+  public BlockModelRegistrationStrategy(Block block) {
 
     this.block = block;
   }